Exit new sighting only if youre on the first screen
